resonanceFit: improve output
Improve the (non-debug) output of the resonance fit. Move some
printing from the 'model' class to the executable itself, because it
is not useful in all cases.
