+++ resonanceFit/model.cc
@@ -248,79 +248,6 @@ rpwa::resonanceFit::model::initMapping(const rpwa::resonanceFit::informationCons
 		}
 	}
 
-	std::ostringstream output;
-	if(_mappingEqualInAllBins) {
-		const size_t idxBin = 0;
-		for(size_t idxWave = 0; idxWave < fitInformation->nrWaves(); ++idxWave) {
-			const rpwa::resonanceFit::information::wave& wave = fitInformation->getWave(idxWave);
-
-			output << "    wave '" << wave.waveName() << "' (index " << idxWave << ") used in" << std::endl;
-			for(size_t idxComponents = 0; idxComponents < _waveComponentChannel[idxBin][idxWave].size(); idxComponents++) {
-				const size_t idxComponent = _waveComponentChannel[idxBin][idxWave][idxComponents].first;
-				const componentConstPtr& component = _components[idxComponent];
-				output << "        component " << idxComponents << ": " << idxComponent << " '" << component->getName() << "'";
-				if(_idxAnchorWave == idxWave and _idxAnchorComponent == idxComponent) {
-					output << " (anchor)";
-				}
-				output << std::endl;
-			}
-		}
-		for(size_t idxComponent = 0; idxComponent < nrComponents; ++idxComponent) {
-			const componentConstPtr& component = _components[idxComponent];
-
-			bool printedHeader = false;
-			for(size_t idxChannel = 0; idxChannel < component->getNrChannels(); ++idxChannel) {
-				const rpwa::resonanceFit::component::channel& channel = component->getChannel(idxChannel);
-				const std::vector<size_t>& bins = channel.getBins();
-				if(std::find(bins.begin(), bins.end(), idxBin) != bins.end()) {
-					if(not printedHeader) {
-						output << "    component '" << component->getName() << "' (index " << idxComponent << ") used in" << std::endl;
-						printedHeader = true;
-					}
-
-					output << "        channel " << idxChannel << ": " << channel.getWaveName() << (channel.isAnchor(idxBin) ? " (anchor)" : "") << std::endl;
-				}
-			}
-		}
-	} else {
-		for(size_t idxBin = 0; idxBin < fitInformation->nrBins(); idxBin++) {
-			output << "    mapping used in bin " << idxBin << std::endl;
-			for(size_t idxWave = 0; idxWave < fitInformation->nrWaves(); ++idxWave) {
-				const rpwa::resonanceFit::information::wave& wave = fitInformation->getWave(idxWave);
-
-				output << "        wave '" << wave.waveName() << "' (index " << idxWave << ") used in" << std::endl;
-				for(size_t idxComponents = 0; idxComponents < _waveComponentChannel[idxBin][idxWave].size(); idxComponents++) {
-					const size_t idxComponent = _waveComponentChannel[idxBin][idxWave][idxComponents].first;
-					const componentConstPtr& component = _components[idxComponent];
-					output << "            component " << idxComponents << ": " << idxComponent << " '" << component->getName() << "'";
-					if(_idxAnchorWave == idxWave and _idxAnchorComponent == idxComponent) {
-						output << " (anchor)";
-					}
-					output << std::endl;
-				}
-			}
-			for(size_t idxComponent = 0; idxComponent < nrComponents; ++idxComponent) {
-				const componentConstPtr& component = _components[idxComponent];
-
-				bool printedHeader = false;
-				for(size_t idxChannel = 0; idxChannel < component->getNrChannels(); ++idxChannel) {
-					const rpwa::resonanceFit::component::channel& channel = component->getChannel(idxChannel);
-					const std::vector<size_t>& bins = channel.getBins();
-					if(std::find(bins.begin(), bins.end(), idxBin) != bins.end()) {
-						if(not printedHeader) {
-							output << "        component '" << component->getName() << "' (index " << idxComponent << ") used in" << std::endl;
-							printedHeader = true;
-						}
-
-						output << "            channel " << idxChannel << ": " << channel.getWaveName() << (channel.isAnchor(idxBin) ? " (anchor)" : "") << std::endl;
-					}
-				}
-			}
-		}
-	}
-	printInfo << fitInformation->nrWaves() << " waves and " << _components.size() << " components in fit model:" << std::endl
-	          << output.str();
-
 	return true;
 }
